requirements.txt
anthropic>=0.49.0
tavily-python>=0.5.0
icalendar>=6.0.0
python-dotenv>=1.0.0
flask>=3.0.0
